fixed wiki page, login page, and modified authenticated route to allow dynamic wagmi load
