+++ src/pages/login/index.tsx
@@ -1,11 +1,22 @@
-import React, { useEffect, useState } from 'react'
+import React, { useContext, useEffect, useState } from 'react'
 import { Box, Container, Heading } from '@chakra-ui/react'
 import Connectors from '@/components/Layout/WalletDrawer/Connectors'
 import { useRouter } from 'next/router'
 import { useAccount } from 'wagmi'
+import { WagmiStatusContext } from '@/components/Wagmi/DynamicWagmiProvider'
+import dynamic from 'next/dynamic'
+import { wagmiNeededRoute } from '@/components/WrapperRoutes/WagmiNeededRoute'
 
 const Login = () => {
   const [isMounted, setIsMounted] = useState(false)
+  const { isWagmiWrapped, setIsWagmiWrapped } = useContext(WagmiStatusContext)
+
+  useEffect(() => {
+    setIsWagmiWrapped(true)
+  }, [setIsWagmiWrapped])
+
+  console.log({ isWagmiWrapped })
+
   const { address: userAddress } = useAccount()
   const router = useRouter()
 
@@ -36,4 +47,6 @@ const Login = () => {
   )
 }
 
-export default Login
+export default dynamic(() => Promise.resolve(wagmiNeededRoute(Login)), {
+  ssr: false,
+})
